refactor(api-websockets-aws): merge with ddb package
