+++ packages/api-event-handler-aws-ddb-os/package.json
@@ -25,7 +25,7 @@
     "@webiny/api-headless-cms-ddb-es": "0.0.0",
     "@webiny/api-opensearch": "0.0.0",
     "@webiny/api-opensearch-aws": "0.0.0",
-    "@webiny/api-websockets-ddb": "0.0.0",
+    "@webiny/api-websockets-aws": "0.0.0",
     "@webiny/db": "0.0.0"
   },
   "peerDependencies": {
