+++ packages/api-event-handler-aws-ddb-os/tsconfig.json
@@ -9,7 +9,7 @@
     { "path": "../api-headless-cms-ddb-es" },
     { "path": "../api-opensearch" },
     { "path": "../api-opensearch-aws" },
-    { "path": "../api-websockets-ddb" },
+    { "path": "../api-websockets-aws" },
     { "path": "../db" }
   ],
   "compilerOptions": {
@@ -33,8 +33,8 @@
       "@webiny/api-opensearch": ["../api-opensearch/src"],
       "@webiny/api-opensearch-aws/*": ["../api-opensearch-aws/src/*"],
       "@webiny/api-opensearch-aws": ["../api-opensearch-aws/src"],
-      "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
-      "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"],
+      "@webiny/api-websockets-aws/*": ["../api-websockets-aws/src/*"],
+      "@webiny/api-websockets-aws": ["../api-websockets-aws/src"],
       "@webiny/db/*": ["../db/src/*"],
       "@webiny/db": ["../db/src"]
     }
