+++ packages/api-event-handler-aws-ddb/package.json
@@ -23,7 +23,7 @@
     "@webiny/api-core-ddb": "0.0.0",
     "@webiny/api-event-handler-aws": "0.0.0",
     "@webiny/api-headless-cms-ddb": "0.0.0",
-    "@webiny/api-websockets-ddb": "0.0.0"
+    "@webiny/api-websockets-aws": "0.0.0"
   },
   "peerDependencies": {
     "@webiny/di": ">=0.2.3"
