+++ packages/api-event-handler-aws-ddb/tsconfig.build.json
@@ -7,7 +7,7 @@
     { "path": "../api-core-ddb/tsconfig.build.json" },
     { "path": "../api-event-handler-aws/tsconfig.build.json" },
     { "path": "../api-headless-cms-ddb/tsconfig.build.json" },
-    { "path": "../api-websockets-ddb/tsconfig.build.json" }
+    { "path": "../api-websockets-aws/tsconfig.build.json" }
   ],
   "compilerOptions": {
     "rootDir": "./src",
@@ -26,8 +26,8 @@
       "@webiny/api-event-handler-aws": ["../api-event-handler-aws/src"],
       "@webiny/api-headless-cms-ddb/*": ["../api-headless-cms-ddb/src/*"],
       "@webiny/api-headless-cms-ddb": ["../api-headless-cms-ddb/src"],
-      "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
-      "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"]
+      "@webiny/api-websockets-aws/*": ["../api-websockets-aws/src/*"],
+      "@webiny/api-websockets-aws": ["../api-websockets-aws/src"]
     }
   }
 }
