+++ packages/api-websockets-aws/package.json
@@ -19,6 +19,8 @@
     "@webiny/api": "0.0.0",
     "@webiny/api-websockets": "0.0.0",
     "@webiny/aws-sdk": "0.0.0",
+    "@webiny/db-dynamodb": "0.0.0",
+    "@webiny/error": "0.0.0",
     "@webiny/event-handler-aws": "0.0.0",
     "@webiny/feature": "0.0.0",
     "@webiny/handler": "0.0.0",
