+++ packages/api-websockets-aws/src/entity.ts
