+++ packages/api-websockets-aws/src/index.ts
@@ -3,6 +3,8 @@ import { AwsWebsocketsTransport } from "~/transport/AwsWebsocketsTransport.js";
 import { WebSocketLambdaHandler } from "~/WebSocketLambdaHandler.js";
 
 export { AwsWebsocketsTransport } from "~/transport/AwsWebsocketsTransport.js";
+export { WebsocketsConnectionRegistry } from "~/WebsocketsConnectionRegistry.js";
+export { WebsocketsDdbFeature } from "~/WebsocketsDdbFeature.js";
 
 export const WebsocketsAwsFeature = createFeature({
     name: "WebsocketsAws",
