+++ packages/api-websockets-aws/tsconfig.json
@@ -5,6 +5,8 @@
     { "path": "../api" },
     { "path": "../api-websockets" },
     { "path": "../aws-sdk" },
+    { "path": "../db-dynamodb" },
+    { "path": "../error" },
     { "path": "../event-handler-aws" },
     { "path": "../feature" },
     { "path": "../handler" },
@@ -29,6 +31,10 @@
       "@webiny/api-websockets": ["../api-websockets/src"],
       "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
       "@webiny/aws-sdk": ["../aws-sdk/src"],
+      "@webiny/db-dynamodb/*": ["../db-dynamodb/src/*"],
+      "@webiny/db-dynamodb": ["../db-dynamodb/src"],
+      "@webiny/error/*": ["../error/src/*"],
+      "@webiny/error": ["../error/src"],
       "@webiny/event-handler-aws/*": ["../event-handler-aws/src/*"],
       "@webiny/event-handler-aws": ["../event-handler-aws/src"],
       "@webiny/feature/*": ["../feature/src/*"],
