--- packages/api-websockets-ddb/__tests__/__api__/presets.js
@@ -1,8 +1,0 @@
-import { resolve } from "path";
-
-export default [
-    {
-        setupFiles: [resolve(import.meta.dirname, "setupFile.js")],
-        setupFilesAfterEnv: [resolve(import.meta.dirname, "setupAfterEnv.js")]
-    }
-];
