--- packages/api-websockets-ddb/__tests__/__api__/setupAfterEnv.js
@@ -1,7 +1,0 @@
-import { resolve } from "path";
-import { setupDynalite } from "@webiny/project-utils/testing/dynalite/index.js";
-
-(async () => {
-    const setupPath = resolve(import.meta.dirname, "../../");
-    await setupDynalite(setupPath);
-})();
