--- packages/api-websockets-ddb/jest-dynalite-config.cjs
@@ -1,2 +1,0 @@
-const { createDynaliteTables } = require("../../dynalite.cjs");
-module.exports = createDynaliteTables();
