--- packages/api-websockets-ddb/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@webiny/api-websockets-ddb",
-  "version": "0.0.0",
-  "type": "module",
-  "keywords": [
-    "@webiny/api-websockets",
-    "storage-operations",
-    "dynamodb",
-    "ddb",
-    "sau:ddb"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/webiny/webiny-js.git"
-  },
-  "description": "DynamoDB storage operations for @webiny/api-websockets.",
-  "author": "Webiny Ltd.",
-  "license": "MIT",
-  "exports": {
-    ".": "./index.js",
-    "./*": "./*"
-  },
-  "dependencies": {
-    "@webiny/api-websockets": "0.0.0",
-    "@webiny/aws-sdk": "0.0.0",
-    "@webiny/db-dynamodb": "0.0.0",
-    "@webiny/error": "0.0.0",
-    "@webiny/feature": "0.0.0"
-  },
-  "devDependencies": {
-    "@webiny/build-tools": "0.0.0",
-    "@webiny/project-utils": "0.0.0",
-    "rimraf": "^6.1.3",
-    "typescript": "^7.0.2",
-    "vitest": "^4.1.10"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "webiny": {
-    "publishFrom": "dist"
-  }
-}
