--- packages/api-websockets-ddb/src/index.ts
@@ -1,2 +1,0 @@
-export { WebsocketsConnectionRegistry } from "./WebsocketsConnectionRegistry.js";
-export { WebsocketsDdbFeature } from "./WebsocketsDdbFeature.js";
