--- packages/api-websockets-ddb/tsconfig.build.json
@@ -1,30 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src"],
-  "references": [
-    { "path": "../api-websockets/tsconfig.build.json" },
-    { "path": "../aws-sdk/tsconfig.build.json" },
-    { "path": "../db-dynamodb/tsconfig.build.json" },
-    { "path": "../error/tsconfig.build.json" },
-    { "path": "../feature/tsconfig.build.json" }
-  ],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": {
-      "~/*": ["./src/*"],
-      "~tests/*": ["./__tests__/*"],
-      "@webiny/api-websockets/*": ["../api-websockets/src/*"],
-      "@webiny/api-websockets": ["../api-websockets/src"],
-      "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
-      "@webiny/aws-sdk": ["../aws-sdk/src"],
-      "@webiny/db-dynamodb/*": ["../db-dynamodb/src/*"],
-      "@webiny/db-dynamodb": ["../db-dynamodb/src"],
-      "@webiny/error/*": ["../error/src/*"],
-      "@webiny/error": ["../error/src"],
-      "@webiny/feature/*": ["../feature/src/*"],
-      "@webiny/feature": ["../feature/src"]
-    }
-  }
-}
