--- packages/api-websockets-ddb/tsconfig.json
@@ -1,30 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__"],
-  "references": [
-    { "path": "../api-websockets" },
-    { "path": "../aws-sdk" },
-    { "path": "../db-dynamodb" },
-    { "path": "../error" },
-    { "path": "../feature" }
-  ],
-  "compilerOptions": {
-    "rootDirs": ["./src", "./__tests__"],
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": {
-      "~/*": ["./src/*"],
-      "~tests/*": ["./__tests__/*"],
-      "@webiny/api-websockets/*": ["../api-websockets/src/*"],
-      "@webiny/api-websockets": ["../api-websockets/src"],
-      "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
-      "@webiny/aws-sdk": ["../aws-sdk/src"],
-      "@webiny/db-dynamodb/*": ["../db-dynamodb/src/*"],
-      "@webiny/db-dynamodb": ["../db-dynamodb/src"],
-      "@webiny/error/*": ["../error/src/*"],
-      "@webiny/error": ["../error/src"],
-      "@webiny/feature/*": ["../feature/src/*"],
-      "@webiny/feature": ["../feature/src"]
-    }
-  }
-}
