--- packages/api-websockets-ddb/vitest.setup.ts
@@ -1,10 +1,0 @@
-import { createTestConfig } from "../../testing";
-
-export default async () => {
-    return createTestConfig({
-        path: import.meta.dirname,
-        vitestConfig: {
-            passWithNoTests: true
-        }
-    });
-};
