+++ packages/project-aws-template/package.json
@@ -45,7 +45,6 @@
     "@webiny/api-website-builder-workflows": "0.0.0",
     "@webiny/api-websockets": "0.0.0",
     "@webiny/api-websockets-aws": "0.0.0",
-    "@webiny/api-websockets-ddb": "0.0.0",
     "@webiny/api-websockets-server": "0.0.0",
     "@webiny/api-websockets-sql": "0.0.0",
     "@webiny/api-workflows": "0.0.0",
@@ -101,7 +100,6 @@
         "@webiny/api-website-builder-workflows",
         "@webiny/api-websockets",
         "@webiny/api-websockets-aws",
-        "@webiny/api-websockets-ddb",
         "@webiny/api-websockets-server",
         "@webiny/api-websockets-sql",
         "@webiny/api-workflows",
